+++ apps/formio-server/server.js
@@ -17,6 +17,7 @@ const os = require('os');
 const license = require('./src/util/license');
 const audit = require('./src/util/audit');
 const vm = require('vm');
+const cors = require('cors');
 const debug = {
   startup: require('debug')('formio:startup')
 };
@@ -81,7 +82,7 @@ module.exports = function(options) {
               else if (config.portalSSOLogout && matched.includes('var ssoLogout =')) {
                 return `var ssoLogout = '${config.portalSSOLogout}';`;
               }
-              else if (matched.includes('var onPremise')) {
+              else if (!process.env.FORMIO_HOSTED && matched.includes('var onPremise')) {
                 return 'var onPremise = true;';
               }
               return matched;
@@ -169,12 +170,15 @@ module.exports = function(options) {
 
   // Status response.
   debug.startup('Attaching middleware: Status');
-  app.get('/status', (req, res) => {
-    res.json({
-      version: packageJson.version,
-      schema: packageJson.schema
-    });
-  });
+  app.get('/status', [
+    cors(),
+    (req, res) => {
+      res.json({
+        version: packageJson.version,
+        schema: packageJson.schema
+      });
+    }
+  ]);
 
   // Load projects and roles.
   debug.startup('Attaching middleware: Project & Roles Loader');
@@ -186,7 +190,7 @@ module.exports = function(options) {
   // CORS Support
   debug.startup('Attaching middleware: CORS');
   var corsMiddleware = require('./src/middleware/corsOptions')(app);
-  var corsRoute = require('cors')(corsMiddleware);
+  var corsRoute = cors(corsMiddleware);
   app.use(function(req, res, next) {
     // If headers already sent, skip cors.
     if (res.headersSent) {
@@ -256,7 +260,7 @@ module.exports = function(options) {
     license.validate(app);
 
     debug.startup('Attaching middleware: Cache');
-    app.formio.formio.cache = _.assign(app.formio.formio.cache, require('./src/cache/cache')(formio));
+    app.formio.formio.cache = _.assign(app.formio.formio.cache, require('./src/cache/cache')(app.formio));
 
     // The formio app sanity endpoint.
     debug.startup('Attaching middleware: Health Check');