+++ apps/formio-server/src/middleware/customCrmAction.js
@@ -11,7 +11,11 @@ module.exports = function(formio) {
 
     if (process.env.hasOwnProperty('CRM')) {
       const modReq = _.cloneDeep(req);
-      modReq.body.name = req.currentProject && req.currentProject.name;
+
+      if (req.currentProject) {
+        modReq.body.name = req.currentProject.name;
+        modReq.body._id = req.currentProject._id;
+      }
 
       const settings = {
         method: 'POST',